Move Projects to a new class
